+++ src/PyF/Internal/QQ.hs
@@ -10,6 +10,9 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE UndecidableInstances #-}
 
+{- | This module uses the python mini language detailed in 'PyF.Internal.PythonSyntax' to build an template haskell expression which represents a 'Formatting.Format'.
+
+-}
 module PyF.Internal.QQ (
   toExp)
 where
@@ -41,6 +44,7 @@ import Data.Proxy
 import GHC.TypeLits
 
 -- Be Careful: empty format string
+-- | Parse a string and return a formatter for it
 toExp:: String -> Q Exp
 toExp s = do
   filename <- loc_filename <$> location
@@ -51,7 +55,7 @@ toExp s = do
         (SourcePos sName _ _) NonEmpty.:| xs = statePos currentState
         in currentState {statePos = (SourcePos sName (mkPos line) (mkPos col)) NonEmpty.:| xs}
 
-  case parse (updateParserState (change_log filename) >> result) filename s of
+  case parse (updateParserState (change_log filename) >> parsePythonFormatString) filename s of
     Left err -> do
 
       if filename == "<interactive>"