+++ ChangeLog.md
@@ -1,4 +1,10 @@
 # Revision history for FormatStringLiteral
+## 0.5.0.0 -- 2018-04-16
+- Support for negative zero
+- Support for 0 modifier
+- Exponential formatter now behaves as python
+- Support for alternate floatting point represenation
+- Lot of documentation
 ## 0.4.0.0 -- 2018-04-13
 
 - Support for grouping option